refactor, reload snippets on file create/delete
