+++ src/snippets/snippet_helper_functions.ts
@@ -1,5 +1,6 @@
+import LatexSuitePlugin from "./../main";
 import { Snippet } from "./snippets";
-import { TFile, TFolder } from "obsidian";
+import { TFile, TFolder, debounce, Notice } from "obsidian";
 
 import { invertedEffects } from "@codemirror/commands";
 import { addMark, removeMark, startSnippet, undidStartSnippet, endSnippet, undidEndSnippet } from "./marker_state_field";
@@ -57,7 +58,7 @@ export function isInFolder(file: TFile, dir: TFolder) {
     let cur = file.parent;
     let cnt = 0;
     
-    while ((!cur.isRoot()) && (cnt < 100)) {
+    while (cur && (!cur.isRoot()) && (cnt < 100)) {
 
         if (cur.path === dir.path) return true;
         
@@ -98,3 +99,57 @@ export const snippetInvertedEffects = invertedEffects.of(tr => {
 
     return effects;
 });
+
+
+
+export async function getSnippetsWithinFolder(folder: TFolder) {
+    let snippets:Snippet[] = [];
+
+    for (const fileOrFolder of folder.children) {
+        if (fileOrFolder instanceof TFile) {
+
+            const content = await this.app.vault.cachedRead(fileOrFolder as TFile);
+
+            try {
+                snippets.push(...getSnippetsFromString(content));
+            }
+            catch (e) {
+                console.log(`Failed to load snippet file ${fileOrFolder.path}:`, e);
+                new Notice(`Failed to load snippet file ${fileOrFolder.name}`);
+            };
+
+        }
+        else {
+            const newSnippets = await this.getSnippetsWithinFolder(fileOrFolder as TFolder);
+            snippets.push(...newSnippets);
+        }
+    }
+
+    return snippets;
+}
+
+
+
+export const debouncedSetSnippetsFromFileOrFolder = debounce(async (plugin: LatexSuitePlugin, path?: string) => {
+
+    if (!path) path = plugin.settings.snippetsFileLocation;
+
+    let snippets:Snippet[];
+    const fileOrFolder = plugin.app.vault.getAbstractFileByPath(path);
+
+
+    if (fileOrFolder instanceof TFolder) {
+        snippets = await getSnippetsWithinFolder(fileOrFolder as TFolder);
+        
+    } else {
+        const content = await plugin.app.vault.cachedRead(fileOrFolder as TFile);
+        snippets = await getSnippetsFromString(content);
+    }
+    
+
+    // Sorting needs to happen after all the snippet files have been parsed
+    sortSnippets(snippets);
+    plugin.snippets = snippets;
+
+    new Notice("Successfully reloaded snippets.", 5000);
+}, 500, true);
